backend/requirements.txt
absl-py==1.2.0
aiosqlite==0.17.0
alembic==1.8.1
anyio==3.6.1
astunparse==1.6.3
asyncpg==0.26.0
atlastk==0.13.2
bcrypt==3.2.2
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
databases==0.6.0
fastapi==0.79.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.47.0
h11==0.13.0
h2==4.1.0
h5py==3.7.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.12.0
iso8601==1.0.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Mako==1.2.1
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
orjson==3.7.11
ormar==0.11.2
packaging==21.3
passlib==1.7.4
Pillow==9.2.0
priority==2.0.0
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pypika-tortoise==0.1.6
python-multipart==0.0.5
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.38
starlette==0.19.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
toml==0.10.2
tortoise==0.1.1
tortoise-orm==0.19.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
websockets==10.3
Werkzeug==2.2.1
wrapt==1.14.1
wsproto==1.1.0
zipp==3.8.1
